Implement a variable polling rate capability.
Implement a polling state which runs through a fast-startup to get a rough
offset correction, a medium-polled acquisition mode to get a rough
frequency correction, and a long-polled maintainance mode (assuming the
time source is good enough) to get a fine frequency correction.

startup mode:		6 five-second polls
acquisition mode:	16 five-minute polls
maintainance mode:	30 minute polls (if the correllation is good enough)

Implement a short failure mode (3 failed polls in a row) which returns
to an aquisition state on recovery, and a long failure mode (10 failed polls
in a row) which returns to a startup state on recovery.

When resetting a regression, also remove any alternate regression further
on down the alt chain (since they won't be staggered any more).