Fix a bug in the mbstats accounting.  m_mbufs was being decremented for normal mbufs converted to clustered mbufs, but in fact it should not be because we decrement it properly when the normal-mbuf-with-cluster is freed.
Change most of the m_getl()'s back to m_get()'s or m_gethdr()'s in order
to guarentee that the mbuf does NOT have a cluster associated with it
since we are going to try to share the old mbuf's cluster later on
(in m_copym()).  This bug was causing the system to run out of mbuf
clusters.

Add a number of assertions for correct operation.