Tokens are recursive in the context of the same thread.  This also means that fast interrupts such as IPIs cannot safely use any token function. Assert this requirement in the implementation.
Add miscellanious other assertions to enforce correct operation.

Remove the 'no giveaway' warning.  This warning was being used to determine
where recursive tokens were being used.  It turns out that they are used
in the mountlist scan code (which has historically been rather buggy).  Now
that we know where they are used and we fixed the recursion release bug in
a prior commit, this printf() can be removed.