Replace three references to /dev/acd0 to /dev/acd0c (for burncd and iso related).
Fix working for the Zwicky study. Capitalize the study. Place name of study
in quotes (according to common style guides).

Remove reference to DragonFly 4.X :)

Add todo:
 - use DragonFly example for cdrecord
  - use DragonFly example for dmesg (for CD-RW)

And change an sect2 header name to not include freebsd4 in name.
(As far as I can see it is not used elsewhere.)