Add counters for recording Token/MPlock contention, this would help in determining the number of times contention has occured in the system.
The contention counters have been made 64-bit quantities because they
are situated within a tight-loop.

KTR tracepoints have been added for marking start and stop of a token's
contention.  New field tr_flags added to struct lwkt_tokref.  By adding
tracepoints in lwkt_chktokens(9),  it gives us interesting data on MP
machines when it indirectly sends a passive IPI to the remote CPU for
gaining ownership of a token.  It would be interesting to see KTR dumps
for a 4-CPU or an 8-CPU system.

Discussed-with: 	Matthew Dillon <dillon@apollo.backplane.com>