When a cpu is stopped due to a panic or the debugger, it can be in virtually any state, including possibly holding a critical section.   IPIQ interrupts must still be processed while we are in this state (even though we could be racing IPIQ processing if we were interrupted at just the wrong time).  In particular, dumping is not likely to work if a panic occurs on a cpu != 0 unless we process the IPIQ on the stopped cpus.  There are simply too many interactions between cpus.
Interrupt threads are LWKT scheduled entities and will generally still not
work during a panic while dumping.  The dumping code expects this.  However,
call splz() anyway.

We may in the future have to allow certain threads to run while dumping.
For example, to allow dumping over the network.  There are various ways this
can be done, such as by masking gd_runqmask or flagging special threads to
be runnable while in a paniced or dumping state.