Add a workaround to make 3COM cardbus cards to propagate the device information to the driver.  This is the first step to get the cards probing and attaching, minus whatever interrupt routing issues.
The proper fix will be addressed by Joerg but we really do require
an interrim solution to the problem.

--
The 575A doesn't have funcregs in memio, so don't claim that it does;
this gets the card to work.  Change obtained from FreeBSD, committed
by Warner Losh to sys/pci/if_xl.c rev. 1.192.