BUF/BIO cleanup 5/99:
Cleanup and document the buffer cache sysctls.  The 'getnewbufrestarts',
'getnewbufcalls', 'bufdefragcnt', 'buffreekvacnt' and 'bufreusecnt' are
now changed to be read-only sysctls.  Group them depending on whether
they are writable or not.

Correct, extend and write documentation for various functions in this
file.

Correct typos in various code comments and adjust nearby style issue.