Add minimal utility that is able to make sense of the per-cpu load accounting struct kinfo_cputime, or previously known as 'cputime'.
It should reside under src/test because it is not complete and only
implements the base logic for those people interested who want to
incorporate this code into top(1) or some other accounting program
of their own.

--
Polish the 'Makefile' while I am around this end of the source tree.