Filesystem journaling.  Reorganize the journal scan for the mountpoint to allow portions of the journal transaction to be written prior to the VOP operation being executed, and accomodate transactional aborts if the VOP fails.   Previously the journal records had to be written only after the VOP succeeded.
Implement UNDO records, including writing out file data that is about to
be overwritten and the mtime prior to a write, rename, or remove.  UNDO
records allow us to create a reversable journal, where we can wind a mirror
forwards or backwards simply by scanning the journal forwards or backwards.