Sync with NetBSD:
PR/23128:

Although the dma descriptors are limited to any 4G address range,
the data is not, using extra area in the dma descriptors to
define a 64 bit address. This is described in appendix B
of the EHCI 1.0 spec from the URL in the file dev/usb/ehci.c.
With a 64 bit address capable controller these upper address
values must be set, the current driver does not do this.