Compile up both the NFS and the TFTP version of the PXE loader and make both available in /boot.  The NFS version obtains all loader-level data (including the kernel binary) from the diskless root mount.  The TFTP version obtains all loader-level data (including the kernel binary) via TFTP.
Do not pass a bogus NFS handle to the kernel (it was previously passing an
all-zero's handle for the TFTP version).

Note that current kernels will not be able to do a NFS root mount from the
tftp version of the loader unless they have BOOTP support compiled in,
because the tftp version of the loader is not able to resolve and pass a
NFS file handle to the kernel.