Rework and expand the algorithms in JSCAN, part 6/?.
Preliminary support for UNDO (-u) on mirrors.  We now have the ability
to roll a mirror forwards or backwards by a specific number of raw journaling
records (-c), or to roll it backwards completely.  However, only basic
remove and write UNDOs currently work, UNDOs for other transaction types
and UNDO's that start mid-transaction are untested.