DragonFly's malloc only guarentees X alignment when X is a power of 2, otherwise we would have to use contigalloc, which is nasty.
Use malloc and adjust the allocation request from a multiple of 512 to the
nearest power of 2 greater or equal to 512 and greater or equal to the
request size.

Once a filesystem is mounted I/O operates on 512 byte boundaries anyway, so
this bug fix only effects programs that operates directly on the block device
such as 'newfs' and 'fsck'.