Synchronise with FreeBSD:
Bump Id where needed.

Constify parameters that do not get touched.

Do something sensible if both -h and -k are given.

If we are asked to print the total number of blocks, do so even if we
have no entries to print (either due to an empty directory or an
error).  This makes the -l and -s options more consistent, like
Solaris and (Debian) Linux.  To make this happen, tweak two
optimizations on the second call to display():

- Don't skip display() altogether, even if list == NULL.
- Don't skip the call to the printfn in display() if we
  need to print the total.

Small clean ups and casts.