Redo the interrupt livelock code.  Simplify the frequency calculations and add a 1-second stabilization requirement before a livelock is removed. Fix a bug that prevented the code from actually limiting the interrupt rate properly.
Greatly simplify ithread_done() and rename it to ithread_unmask().  It is
no longer responsible for putting the interrupt thread to sleep.

There is one issue left and that is the livelock code must re-enable the
interrupt at the limit frequency in order to detect when the actual
interrupt rate falls below the low water mark.  In a livelocked situation
the actual interrupt rate will be about 2x the livelock limit because
both the systimer interrupt and the livelocked interrupt will be running.